+++ device-minion-runtime/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [SW-15697] Replacing threshold_data with threshold value in trace buffer full notification.
 - [SW-16430] Optimizing the kernel launch path to process command fields before waiting on idle slot search.
 - [SW-16452] Generate pre-release packages by default and full versions in tags pipelines
+- [SW-15063] Moving out non-generic C flags from toolchain config to consumer projects
 ### Deprecated
 ### Removed
 ### Fixed